% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kantarPal.R
\docType{data}
\name{kantar_colors}
\alias{kantar_colors}
\title{Kantar Corporate Colors}
\format{An object of class \code{character} of length 19.}
\usage{
kantar_colors
}
\description{
Kantar Corporate Colors
}
\keyword{datasets}
