% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kantarPal.R
\name{kantar_cols}
\alias{kantar_cols}
\title{Function to extract kantar colors as hex codes}
\usage{
kantar_cols(...)
}
\arguments{
\item{...}{Character names of kantar_colors}
}
\description{
Function to extract kantar colors as hex codes
}
