% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kantarPal.R
\name{kantar_pal}
\alias{kantar_pal}
\title{Return function to interpolate a kantar color palette}
\usage{
kantar_pal(palette = "main", reverse = FALSE, expand_by = 0, ...)
}
\arguments{
\item{palette}{Character name of palette in kantar_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a kantar color palette
}
