% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kantarPal.R
\docType{data}
\name{kantar_palettes}
\alias{kantar_palettes}
\title{Create Kantar Pallets}
\format{An object of class \code{list} of length 9.}
\usage{
kantar_palettes
}
\description{
Create Kantar Pallets
}
\keyword{datasets}
