% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kantarPal.R
\name{scale_color_kantar}
\alias{scale_color_kantar}
\title{Color scale constructor for kantar colors}
\usage{
scale_color_kantar(palette = "main", discrete = TRUE, reverse = FALSE,
  expand_by = 0, ...)
}
\arguments{
\item{palette}{Character name of palette in kantar_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for kantar colors
}
